requirements.txt
pyspark==3.5.0
pandas
cassandra-driver
kafka-python
scikit-learn
